+++ src/main.rs
@@ -15,10 +15,11 @@ pub mod client;
 pub mod session;
 pub mod config;
 pub mod dns;
+pub mod util;
 
 use std::net::Ipv4Addr;
 use std::thread;
-use std::io::Read;
+use std::io::{self, Read};
 use std::fs::File;
 
 use clap::{App, Arg};
@@ -31,7 +32,7 @@ fn main() {
     env_logger::init().unwrap();
 
     let matches = App::new("Telepipe")
-        .version("1.1.0")
+        .version("1.1.1")
         .author("Eidolon (@HybridEidolon)")
         .about("A proxy for Phantasy Star Online games for the GameCube.")
         .arg(Arg::with_name("config")
@@ -61,7 +62,17 @@ fn main() {
     if config.use_dns {
         use std::str::FromStr;
 
-        let sock = DnsSocket::bind((config.bind_addr.as_str(), 53)).unwrap();
+        let sock = match DnsSocket::bind((config.bind_addr.as_str(), 53)) {
+            Ok(v) => v,
+            Err(e) => {
+                println!("An error binding the DNS server occurred: {}", e);
+                if e.kind() == io::ErrorKind::AddrNotAvailable {
+                    println!("If you are trying to play in Dolphin, it is likely that you have not initialized the TAP adapter yet. Please make sure to install and setup your TAP adapter and wait until the 'Connecting to DNS Server' prompt in-game before running this program.");
+                }
+                util::pause().unwrap();
+                return
+            }
+        };
         let proxy_server_addr = Ipv4Addr::from_str(config.proxy_server_addr.as_str()).unwrap();
         thread::spawn(move || {
             dns::dns_thread(sock, proxy_server_addr);
@@ -69,7 +80,17 @@ fn main() {
     }
 
     for (local, server) in config.listen_ports {
-        spawn_new_session((config.bind_addr.as_str(), local), (config.server_addr.as_str(), server), config.proxy_server_addr.clone(), config.bind_addr.clone(), false);
+        match spawn_new_session((config.bind_addr.as_str(), local), (config.server_addr.as_str(), server), config.proxy_server_addr.clone(), config.bind_addr.clone(), false) {
+            Err(e) => {
+                println!("An error occurred: {}", e);
+                match e.kind() {
+                    _ => println!("{:?}", e)
+                }
+                util::pause().unwrap();
+                return
+            },
+            _ => ()
+        }
     }
 
     loop {