+++ src/session.rs
@@ -1,5 +1,7 @@
 //! Spawns a thread that wraps a TCP Listener socket with a new destination.
 
+use std::io;
+use std::fmt::Debug;
 use std::net::{TcpListener, ToSocketAddrs, SocketAddr};
 use std::thread;
 
@@ -7,16 +9,16 @@ use client::{Session, Run};
 
 /// Spawn a session listener thread on the addresses given. If the local port is 0, a port will be
 /// picked. The port of the local listener is returned.
-pub fn spawn_new_session<A: ToSocketAddrs, B: ToSocketAddrs>(local_addr: A, server_addr: B, local_server_addr: String, local_bind_addr: String, kill_on_connect: bool) -> u16 {
-    let tcp_listener: TcpListener = TcpListener::bind(local_addr).unwrap();
-    let port = tcp_listener.local_addr().unwrap().port();
-    let server_socket_addr = server_addr.to_socket_addrs().unwrap().next().unwrap();
+pub fn spawn_new_session<A: ToSocketAddrs, B: ToSocketAddrs + Debug>(local_addr: A, server_addr: B, local_server_addr: String, local_bind_addr: String, kill_on_connect: bool) -> io::Result<u16> {
+    let tcp_listener: TcpListener = TcpListener::bind(local_addr)?;
+    let port = tcp_listener.local_addr()?.port();
+    let server_socket_addr = server_addr.to_socket_addrs()?.next().ok_or(io::Error::new(io::ErrorKind::AddrNotAvailable, format!("Address {:?} invalid", server_addr)))?;
 
     thread::spawn(move|| {
         session_thread(tcp_listener, server_socket_addr, local_server_addr, local_bind_addr, kill_on_connect);
     });
 
-    port
+    Ok(port)
 }
 
 fn session_thread(listener: TcpListener, server_addr: SocketAddr, local_server_addr: String, local_bind_addr: String, kill_on_connect: bool) {